Adding runtimes is no longer part of the facet install action, it is a separate step.
