+++ plugins/com.google.cloud.tools.eclipse.appengine.facets/src/com/google/cloud/tools/eclipse/appengine/facets/FacetInstallDelegate.java
@@ -60,144 +60,8 @@ public class FacetInstallDelegate implements IDelegate {
                       IProgressMonitor monitor) throws CoreException {
     if (!MavenUtils.hasMavenNature(project)) { // Maven handles classpath in maven projects.
       addAppEngineJarsToClasspath(project, monitor);
-      IFacetedProject facetedProject = ProjectFacetsManager.create(project);
-      installAppEngineRuntime(facetedProject, monitor);
       createConfigFiles(project, monitor);
     }
-
-  }
-
-  /**
-   * If App Engine runtimes exist in the workspace, add them to the list of targeted runtimes
-   * of <code>project</code>, otherwise create a new App Engine runtime and add it to the list
-   * of targeted runtimes.
-   *
-   * @param project the workspace faceted project
-   * @param monitor the progress monitor
-   * @throws CoreException
-   */
-  public static void installAppEngineRuntime(IFacetedProject project, IProgressMonitor monitor)
-      throws CoreException {
-    // TODO what happens if the project already has App Engine runtime?
-    Set<IProjectFacetVersion> facets = new HashSet<>();
-    facets.add(WebFacetUtils.WEB_25);
-    Set<IRuntime> runtimes = RuntimeManager.getRuntimes(facets);
-    final IFacetedProjectWorkingCopy fpwc = project.createWorkingCopy();
-    fpwc.setTargetedRuntimes(runtimes);
-    org.eclipse.wst.server.core.IRuntime[] appEngineRuntimes = getAppEngineRuntime();
-
-    if (appEngineRuntimes.length > 0) {
-      IRuntime appEngineFacetRuntime = null;
-      for(int index = 0; index < appEngineRuntimes.length; index++) {
-        appEngineFacetRuntime = FacetUtil.getRuntime(appEngineRuntimes[index]);
-        fpwc.addTargetedRuntime(appEngineFacetRuntime);
-      }
-      fpwc.setPrimaryRuntime(appEngineFacetRuntime);
-    } else { // Create a new App Engine runtime
-      IRuntimeType appEngineRuntimeType =
-          ServerCore.findRuntimeType(AppEngineStandardFacet.DEFAULT_RUNTIME_ID);
-      if (appEngineRuntimeType == null) {
-        throw new NullPointerException("Could not find " + AppEngineStandardFacet.DEFAULT_RUNTIME_NAME + " runtime type");
-      }
-
-      IRuntimeWorkingCopy appEngineRuntimeWorkingCopy
-          = appEngineRuntimeType.createRuntime(null, monitor);
-      CloudSdk cloudSdk = new CloudSdk.Builder().build();
-      if (cloudSdk != null) {
-        java.nio.file.Path sdkLocation = cloudSdk.getJavaAppEngineSdkPath();
-        if (sdkLocation != null) {
-          IPath sdkPath = Path.fromOSString(sdkLocation.toAbsolutePath().toString());
-          appEngineRuntimeWorkingCopy.setLocation(sdkPath);
-        }
-      }
-
-      org.eclipse.wst.server.core.IRuntime appEngineServerRuntime =
-          appEngineRuntimeWorkingCopy.save(true, monitor);
-      IRuntime appEngineFacetRuntime = FacetUtil.getRuntime(appEngineServerRuntime);
-      if (appEngineFacetRuntime == null) {
-        throw new NullPointerException("Could not locate App Engine facet runtime");
-      }
-
-      fpwc.addTargetedRuntime(appEngineFacetRuntime);
-      fpwc.setPrimaryRuntime(appEngineFacetRuntime);
-    }
-
-    // This is to prevent the error "Cannot modify faceted project from within a facet delegate"
-    Job commitChanges = new Job("Add App Engine runtime to " + fpwc.getProjectName()) {
-
-      @Override
-      protected IStatus run(IProgressMonitor monitor) {
-        try {
-          fpwc.commitChanges(monitor);
-        } catch (CoreException e) {
-          return e.getStatus();
-        }
-        return Status.OK_STATUS;
-      }
-
-    };
-    commitChanges.schedule();
-  }
-
-  /**
-   * Checks to see if <code>facetedProject</code> has the App Engine facet installed. If not, it installs
-   * the App Engine facet.
-   *
-   * @param facetedProject the workspace faceted project
-   * @param monitor the progress monitor
-   * @throws CoreException
-   */
-  public static void installAppEngineFacet(IFacetedProject facetedProject, boolean installDependentFacets, IProgressMonitor monitor)
-      throws CoreException {
-    // Install required App Engine facets Java 1.7 and Dynamic Web Module 2.5
-    if (installDependentFacets) {
-      installJavaFacet(facetedProject, monitor);
-      installWebFacet(facetedProject, monitor);
-    }
-
-    IProjectFacet appEngineFacet = ProjectFacetsManager.getProjectFacet(AppEngineStandardFacet.ID);
-    IProjectFacetVersion appEngineFacetVersion = appEngineFacet.getVersion(AppEngineStandardFacet.VERSION);
-
-    if (!facetedProject.hasProjectFacet(appEngineFacet)) {
-      IFacetedProjectWorkingCopy workingCopy = facetedProject.createWorkingCopy();
-      workingCopy.addProjectFacet(appEngineFacetVersion);
-      workingCopy.commitChanges(monitor);
-    }
-  }
-
-  /**
-   * Installs Java 1.7 facet if it doesn't already exits in <code>factedProject</code>
-   */
-  private static void installJavaFacet(IFacetedProject facetedProject, IProgressMonitor monitor)
-      throws CoreException {
-    if (facetedProject.hasProjectFacet(JavaFacet.VERSION_1_7)) {
-      return;
-    }
-
-    JavaFacetInstallConfig javaConfig = new JavaFacetInstallConfig();
-    List<IPath> sourcePaths = new ArrayList<>();
-    sourcePaths.add(new Path("src/main/java"));
-    sourcePaths.add(new Path("src/test/java"));
-    javaConfig.setSourceFolders(sourcePaths);
-    // TODO: when do we use IFacetedProject#installProjectFacet vs IFacetedProjectWorkingCopy#addProjectFacet
-    facetedProject.installProjectFacet(JavaFacet.VERSION_1_7, javaConfig, monitor);
-  }
-
-  /**
-   * Installs Dynamic Web Module 2.5 facet if it doesn't already exits in <code>factedProject</code>
-   */
-  private static void installWebFacet(IFacetedProject facetedProject, IProgressMonitor monitor)
-      throws CoreException {
-    if (facetedProject.hasProjectFacet(WebFacetUtils.WEB_25)) {
-      return;
-    }
-
-    IDataModel webModel = DataModelFactory.createDataModel(new WebFacetInstallDataModelProvider());
-    webModel.setBooleanProperty(IJ2EEModuleFacetInstallDataModelProperties.ADD_TO_EAR, false);
-    webModel.setBooleanProperty(IJ2EEFacetInstallDataModelProperties.GENERATE_DD, false);
-    webModel.setBooleanProperty(IWebFacetInstallDataModelProperties.INSTALL_WEB_LIBRARY, false);
-    webModel.setStringProperty(IWebFacetInstallDataModelProperties.CONFIG_FOLDER, "src/main/webapp");
-    facetedProject.installProjectFacet(WebFacetUtils.WEB_25, webModel, monitor);
   }
 
   /**
@@ -226,22 +90,6 @@ public class FacetInstallDelegate implements IDelegate {
     System.arraycopy(rawClasspath, 0, newClasspath, 0, rawClasspath.length);
     newClasspath[newClasspath.length - 1] = appEngineContainer;
     javaProject.setRawClasspath(newClasspath, monitor);
-  }
-
-  // TODO: find a more general form of this method
-  private static org.eclipse.wst.server.core.IRuntime[] getAppEngineRuntime() {
-    org.eclipse.wst.server.core.IRuntime[] allRuntimes = ServerCore.getRuntimes();
-    List<org.eclipse.wst.server.core.IRuntime> appEngineRuntimes = new ArrayList<>();
-
-    for (int i = 0; i < allRuntimes.length; i++) {
-      if (allRuntimes[i].getRuntimeType().getId().equals(AppEngineStandardFacet.DEFAULT_RUNTIME_ID)) {
-        appEngineRuntimes.add(allRuntimes[i]);
-      }
-    }
-
-    org.eclipse.wst.server.core.IRuntime[] appEngineRuntimesArray =
-        new org.eclipse.wst.server.core.IRuntime[appEngineRuntimes.size()];
-    return appEngineRuntimes.toArray(appEngineRuntimesArray);
   }
 
   /**
