+++ plugins/com.google.cloud.tools.eclipse.appengine.newproject/META-INF/MANIFEST.MF
@@ -20,6 +20,7 @@ Require-Bundle: com.google.cloud.tools.app.lib;bundle-version="0.1.0",
  com.google.cloud.tools.eclipse.sdk.ui;bundle-version="0.1.0"
 Import-Package: com.google.cloud.tools.eclipse.appengine.facets,
  org.eclipse.core.runtime;version="3.5.0",
+ org.eclipse.core.runtime.jobs,
  org.eclipse.jdt.core,
  org.eclipse.jst.j2ee.project.facet,
  org.eclipse.jst.server.core,
